+++ javaparser-symbol-solver-testing/src/test/resources/TypeResolutionWithSameNameTest/src1/extends_duplicate/A.java
@@ -1,4 +1,4 @@
-package testresource;
+package extends_duplicate;
 
 public class A extends DuplicateTypeName {
     class DuplicateTypeName extends A {
