+++ javaparser-symbol-solver-testing/src/test/resources/TypeResolutionWithSameNameTest/src1/extends_duplicate/DuplicateTypeName.java
@@ -1,4 +1,4 @@
-package testresource;
+package extends_duplicate;
 
 public abstract class DuplicateTypeName {
