+++ javaparser-symbol-solver-testing/src/test/resources/TypeResolutionWithSameNameTest/src2/implements_duplicate/A.java
@@ -1,4 +1,4 @@
-package testresource;
+package implements_duplicate;
 
 public class A implements DuplicateTypeName {
     class DuplicateTypeName extends A {
