+++ javaparser-symbol-solver-testing/src/test/resources/TypeResolutionWithSameNameTest/src2/implements_duplicate/DuplicateTypeName.java
@@ -1,4 +1,4 @@
-package testresource;
+package implements_duplicate;
 
 public interface DuplicateTypeName {
